consolidate with CORS middleware
